+++ .env
@@ -1,0 +1,3 @@
+VERCEL_URL = https://vercel.com/marwans-projects-89c0b6a9/typo-tester/Dc9WEk2uRA7MTqYhEksc2sfpRPGk
+
+serverPort = http://localhost:5000
