+++ src/api/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "api",
+  "version": "1.0.0",
+  "description": "back-end for typotester",
+  "main": "server.js",
+  "scripts": {
+    "start": "node server.cjs",
+    "test": "jest"
+  },
+  "dependencies": {
+    "@vercel/node": "^3.0.12",
+    "axios": "^1.6.2",
+    "body-parser": "^1.20.2",
+    "cors": "^2.8.5",
+    "create-vite": "^5.1.0",
+    "express": "^4.18.2",
+    "https": "^1.0.0",
+    "random-words": "^2.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "socket.io": "^4.7.2",
+    "socket.io-client": "^4.7.2"
+  },
+  "author": "jagmw",
+  "license": "ISC"
+}
