+++ src/api/vercel.json
@@ -1,0 +1,11 @@
+{
+  "routes": [
+    {
+      "src": "/(.*)",
+      "dest": "/server.js",
+      "methods": ["GET","HEAD","PUT","PATCH","POST","DELETE"],
+      "headers": { "Access-Control-Allow-Origrn": "*" }
+    }
+  ],
+  "builds": [{ "src": "/*.js", "use": "@vercel/node" }]
+}
